Fix unnecessary cast

+++ src/loader.rs
@@ -272,7 +272,7 @@ impl<St: Stream> Stream for ResultStream<St> {
     type Item = St::Item;
 
     fn size_hint(&self) -> (usize, Option<usize>) {
-        (self.total as usize, None)
+        (self.total, None)
     }
 
     fn poll_next(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Option<Self::Item>> {
